[thread-helper] move thread_helper to host/
`ThreadHelper` is a class with many responsibilities and dependencies.
It's currently in `utils/` which contians a bunch of low level
components. Moving `ThreadHelper` to `host` can make the `utils` module
more self-contained and more friendly for build systems.
