+++ tests/gtest/fake_posix_platform.cpp
@@ -31,6 +31,7 @@
 #include <assert.h>
 
 #include "openthread/openthread-system.h"
+#include "openthread/platform/dns.h"
 
 otPlatResetReason        gPlatResetReason = OT_PLAT_RESET_REASON_POWER_ON;
 static ot::FakePlatform *sFakePlatform;
@@ -56,6 +57,32 @@ void otSysCountInfraNetifAddresses(otSysInfraNetIfAddressCounters *)
 const char *otSysGetInfraNetifName(void)
 {
     return nullptr;
+}
+
+bool otPlatDnsIsUpstreamQueryAvailable(otInstance *aInstance)
+{
+    OT_UNUSED_VARIABLE(aInstance);
+    return true;
+}
+
+void otPlatDnsStartUpstreamQuery(otInstance *aInstance, otPlatDnsUpstreamQuery *aTxn, const otMessage *aQuery)
+{
+    OT_UNUSED_VARIABLE(aInstance);
+    OT_UNUSED_VARIABLE(aTxn);
+    OT_UNUSED_VARIABLE(aQuery);
+}
+
+void otPlatDnsCancelUpstreamQuery(otInstance *aInstance, otPlatDnsUpstreamQuery *aTxn)
+{
+    OT_UNUSED_VARIABLE(aInstance);
+    OT_UNUSED_VARIABLE(aTxn);
+}
+
+void otPlatDnsUpstreamQueryDone(otInstance *aInstance, otPlatDnsUpstreamQuery *aTxn, otMessage *aResponse)
+{
+    OT_UNUSED_VARIABLE(aInstance);
+    OT_UNUSED_VARIABLE(aTxn);
+    OT_UNUSED_VARIABLE(aResponse);
 }
 
 otInstance *otSysInit(otPlatformConfig *aPlatformConfig)
